Refactor Firebase imports and add test environment API
- Updated Firebase admin imports in login, register, and dashboard API files to use the new chunk.
- Introduced a new API endpoint for testing environment variables, returning relevant Firebase configuration details.
- Updated Vercel configuration to include the new test environment route.
- Simplified vercel.json by removing unnecessary function settings and added build command and output directory.
