+++ .vercel/output/_functions/pages/api/auth/login.astro.mjs
@@ -1,8 +1,13 @@
-import { a as adminAuth, b as adminDb } from '../../../chunks/firebase-admin_DsrFsXfx.mjs';
+import { a as adminAuth, b as adminDb } from '../../../chunks/firebase-admin_D9DK7rPw.mjs';
 export { renderers } from '../../../renderers.mjs';
 
 const prerender = false;
 const POST = async ({ request, cookies }) => {
+  console.log("Login API called at:", (/* @__PURE__ */ new Date()).toISOString());
+  console.log("Environment check:", {
+    NODE_ENV: process.env.NODE_ENV,
+    FIREBASE_PROJECT_ID: "✓" 
+  });
   try {
     const body = await request.json();
     const { email, password } = body;
