+++ .vercel/output/config.json
@@ -36,6 +36,10 @@
 			"dest": "_render"
 		},
 		{
+			"src": "^/api/test-env/?$",
+			"dest": "_render"
+		},
+		{
 			"src": "^/contact/?$",
 			"dest": "_render"
 		},
